test(key-gen): cleanup tests move integration -> unit
