+++ oprf-key-gen/Cargo.toml
@@ -75,3 +75,4 @@ zeroize = { workspace = true, features = ["derive"] }
 [dev-dependencies]
 axum-test = { workspace = true, features = ["ws"] }
 oprf-test-utils = { package = "taceo-oprf-test-utils", path = "../oprf-test-utils", version = "0.10" }
+parking_lot = { workspace = true }
