+++ oprf-key-gen/src/lib.rs
@@ -52,6 +52,9 @@ pub mod config;
 pub mod metrics;
 pub(crate) mod services;
 
+#[cfg(test)]
+mod tests;
+
 pub use nodes_common::Environment;
 pub use nodes_common::StartedServices;
 use secrecy::ExposeSecret;
