+++ oprf-key-gen/src/services/key_event_watcher/tests.rs
@@ -1,14 +1,7 @@
-use std::{
-    collections::HashMap,
-    str::FromStr,
-    sync::{Arc, Mutex},
-    time::Duration,
-};
+use std::{str::FromStr, time::Duration};
 
 use crate::{
-    secret_manager::{
-        KeyGenIntermediateValues, SecretManager, SecretManagerError, SecretManagerService,
-    },
+    secret_manager::{SecretManager, SecretManagerError, SecretManagerService},
     services::{
         key_event_watcher::{
             TransactionError, handle_abort, handle_delete, handle_not_enough_producers,
@@ -16,17 +9,13 @@ use crate::{
         secret_gen::DLogSecretGenService,
         transaction_handler::{TransactionHandler, TransactionHandlerArgs},
     },
+    tests::keygen_test_secret_manager::TestKeyGenSecretManager,
 };
 use alloy::{network::EthereumWallet, primitives::U160, signers::local::PrivateKeySigner};
-use ark_serialize::{CanonicalDeserialize, CanonicalSerialize};
-use async_trait::async_trait;
 use groth16_material::circom::{CircomGroth16Material, CircomGroth16MaterialBuilder};
 use nodes_common::{Environment, web3::RpcProviderBuilder};
 use oprf_core::ddlog_equality::shamir::DLogShareShamir;
-use oprf_test_utils::{
-    DeploySetup, OPRF_PEER_PRIVATE_KEY_0, PEER_ADDRESSES, TestSetup,
-    test_secret_manager::TestSecretManager,
-};
+use oprf_test_utils::{DeploySetup, OPRF_PEER_PRIVATE_KEY_0, PEER_ADDRESSES, TestSetup};
 use oprf_types::{
     OprfKeyId, ShareEpoch,
     chain::{OprfKeyRegistry, RevertError, Verifier::VerifierErrors},
@@ -35,175 +24,6 @@ use oprf_types::{
 
 const INVALID_PROOF_KEY: usize = 43;
 const WRONG_ROUND_LOAD_PEER_PUBLIC_KEYS: usize = 44;
-
-struct TestKeyGenSecretManager {
-    base: Arc<TestSecretManager>,
-    keygen_intermediates: Mutex<HashMap<(OprfKeyId, ShareEpoch), Vec<u8>>>,
-    pending_shares: Mutex<HashMap<(OprfKeyId, ShareEpoch), DLogShareShamir>>,
-}
-
-impl TestKeyGenSecretManager {
-    fn new(base: Arc<TestSecretManager>) -> Self {
-        Self {
-            base,
-            keygen_intermediates: Mutex::new(HashMap::new()),
-            pending_shares: Mutex::new(HashMap::new()),
-        }
-    }
-
-    fn serialize_intermediates(
-        intermediate: &KeyGenIntermediateValues,
-    ) -> Result<Vec<u8>, SecretManagerError> {
-        let mut bytes = Vec::with_capacity(intermediate.uncompressed_size());
-        intermediate
-            .serialize_uncompressed(&mut bytes)
-            .map_err(|error| SecretManagerError::Internal(eyre::eyre!(error)))?;
-        Ok(bytes)
-    }
-
-    fn deserialize_intermediates(
-        bytes: &[u8],
-    ) -> Result<KeyGenIntermediateValues, SecretManagerError> {
-        KeyGenIntermediateValues::deserialize_uncompressed(bytes)
-            .map_err(|error| SecretManagerError::Internal(eyre::eyre!(error)))
-    }
-}
-
-#[async_trait]
-impl SecretManager for TestKeyGenSecretManager {
-    async fn store_wallet_address(&self, address: String) -> Result<(), SecretManagerError> {
-        self.base
-            .store_wallet_address(address)
-            .await
-            .map_err(Into::into)
-    }
-
-    async fn get_share_by_epoch(
-        &self,
-        oprf_key_id: OprfKeyId,
-        generated_epoch: ShareEpoch,
-    ) -> Result<Option<DLogShareShamir>, SecretManagerError> {
-        self.base
-            .get_share_by_epoch(oprf_key_id, generated_epoch)
-            .await
-            .map_err(Into::into)
-    }
-
-    async fn delete_oprf_key_material(
-        &self,
-        oprf_key_id: OprfKeyId,
-    ) -> Result<(), SecretManagerError> {
-        self.base.remove_key_material(oprf_key_id);
-        self.keygen_intermediates
-            .lock()
-            .expect("lock not poisoned")
-            .retain(|(key_id, _), _| *key_id != oprf_key_id);
-        self.pending_shares
-            .lock()
-            .expect("lock not poisoned")
-            .retain(|(key_id, _), _| *key_id != oprf_key_id);
-        Ok(())
-    }
-
-    async fn abort_keygen(&self, oprf_key_id: OprfKeyId) -> Result<(), SecretManagerError> {
-        self.keygen_intermediates
-            .lock()
-            .expect("lock not poisoned")
-            .retain(|(key_id, _), _| *key_id != oprf_key_id);
-        self.pending_shares
-            .lock()
-            .expect("lock not poisoned")
-            .retain(|(key_id, _), _| *key_id != oprf_key_id);
-        Ok(())
-    }
-
-    async fn try_store_keygen_intermediates(
-        &self,
-        oprf_key_id: OprfKeyId,
-        pending_epoch: ShareEpoch,
-        intermediate: KeyGenIntermediateValues,
-    ) -> Result<KeyGenIntermediateValues, SecretManagerError> {
-        let mut intermediates = self.keygen_intermediates.lock().expect("lock not poisoned");
-        let serialized = Self::serialize_intermediates(&intermediate)?;
-        Self::deserialize_intermediates(
-            intermediates
-                .entry((oprf_key_id, pending_epoch))
-                .or_insert(serialized),
-        )
-    }
-
-    async fn fetch_keygen_intermediates(
-        &self,
-        oprf_key_id: OprfKeyId,
-        pending_epoch: ShareEpoch,
-    ) -> Result<Option<KeyGenIntermediateValues>, SecretManagerError> {
-        self.keygen_intermediates
-            .lock()
-            .expect("lock not poisoned")
-            .get(&(oprf_key_id, pending_epoch))
-            .map(|bytes| Self::deserialize_intermediates(bytes))
-            .transpose()
-    }
-
-    async fn store_pending_dlog_share(
-        &self,
-        oprf_key_id: OprfKeyId,
-        pending_epoch: ShareEpoch,
-        share: DLogShareShamir,
-    ) -> Result<(), SecretManagerError> {
-        if !self
-            .keygen_intermediates
-            .lock()
-            .expect("lock not poisoned")
-            .contains_key(&(oprf_key_id, pending_epoch))
-        {
-            return Err(SecretManagerError::MissingIntermediates(
-                oprf_key_id,
-                pending_epoch,
-            ));
-        }
-        self.pending_shares
-            .lock()
-            .expect("lock not poisoned")
-            .insert((oprf_key_id, pending_epoch), share);
-        Ok(())
-    }
-
-    async fn confirm_dlog_share(
-        &self,
-        oprf_key_id: OprfKeyId,
-        epoch: ShareEpoch,
-        public_key: OprfPublicKey,
-    ) -> Result<(), SecretManagerError> {
-        let share = self
-            .pending_shares
-            .lock()
-            .expect("lock not poisoned")
-            .remove(&(oprf_key_id, epoch))
-            .ok_or(SecretManagerError::MissingIntermediates(oprf_key_id, epoch))?;
-
-        if let Some(existing) = self.base.get_key_material(oprf_key_id)
-            && existing.epoch() >= epoch
-        {
-            return Err(SecretManagerError::RefusingToRollbackEpoch);
-        }
-
-        self.base.insert_key_material(
-            oprf_key_id,
-            oprf_types::crypto::OprfKeyMaterial::new(share, public_key, epoch),
-        );
-        self.keygen_intermediates
-            .lock()
-            .expect("lock not poisoned")
-            .retain(|(key_id, _), _| *key_id != oprf_key_id);
-        self.pending_shares
-            .lock()
-            .expect("lock not poisoned")
-            .retain(|(key_id, _), _| *key_id != oprf_key_id);
-        Ok(())
-    }
-}
-
 fn key_gen_material(deploy_setup: DeploySetup) -> CircomGroth16Material {
     CircomGroth16MaterialBuilder::new()
         .bbf_inv()
@@ -212,10 +32,8 @@ fn key_gen_material(deploy_setup: DeploySetup) -> CircomGroth16Material {
         .expect("Can build key_gen_material")
 }
 
-fn test_secret_manager() -> (Arc<TestSecretManager>, Arc<TestKeyGenSecretManager>) {
-    let base = Arc::new(TestSecretManager::new(OPRF_PEER_PRIVATE_KEY_0));
-    let secret_manager = Arc::new(TestKeyGenSecretManager::new(Arc::clone(&base)));
-    (base, secret_manager)
+fn test_secret_manager() -> TestKeyGenSecretManager {
+    TestKeyGenSecretManager::new(OPRF_PEER_PRIVATE_KEY_0)
 }
 
 fn random_share() -> DLogShareShamir {
@@ -257,9 +75,9 @@ async fn test_config(setup: &TestSetup) -> (CircomGroth16Material, TransactionHa
 async fn test_send_invalid_proof() -> eyre::Result<()> {
     let setup = TestSetup::new(DeploySetup::TwoThree).await?;
     let (key_gen_material, transaction_handler) = test_config(&setup).await;
-    let (_, secret_manager) = test_secret_manager();
-    let secret_manager: SecretManagerService = secret_manager;
-    let secret_gen = DLogSecretGenService::init(key_gen_material, secret_manager);
+    let secret_manager = test_secret_manager();
+    let secret_manager_service: SecretManagerService = secret_manager.service();
+    let secret_gen = DLogSecretGenService::init(key_gen_material, secret_manager_service);
     let key_id = U160::from(INVALID_PROOF_KEY);
     secret_gen
         .key_gen_round1(key_id.into(), ShareEpoch::default(), 2)
@@ -282,8 +100,8 @@ async fn test_send_invalid_proof() -> eyre::Result<()> {
 
 #[tokio::test]
 async fn test_delete() -> eyre::Result<()> {
-    let (base, secret_manager) = test_secret_manager();
-    let secret_manager_service: SecretManagerService = secret_manager.clone();
+    let secret_manager = test_secret_manager();
+    let secret_manager_service: SecretManagerService = secret_manager.service();
     let secret_gen = DLogSecretGenService::init(
         key_gen_material(DeploySetup::TwoThree),
         secret_manager_service,
@@ -292,7 +110,7 @@ async fn test_delete() -> eyre::Result<()> {
     let oprf_key_id = OprfKeyId::new(U160::from(42));
     let confirmed_epoch = ShareEpoch::default();
     let pending_epoch = confirmed_epoch.next();
-    base.add_random_key_material_with_id_epoch(
+    secret_manager.add_random_key_material_with_id_epoch(
         oprf_key_id,
         confirmed_epoch,
         &mut rand::thread_rng(),
@@ -351,8 +169,8 @@ async fn test_delete() -> eyre::Result<()> {
 async fn test_round2_in_wrong_round_during_load_public_keys() -> eyre::Result<()> {
     let setup = TestSetup::new(DeploySetup::TwoThree).await?;
     let (key_gen_material, transaction_handler) = test_config(&setup).await;
-    let (_, secret_manager) = test_secret_manager();
-    let secret_manager_service: SecretManagerService = secret_manager.clone();
+    let secret_manager = test_secret_manager();
+    let secret_manager_service: SecretManagerService = secret_manager.service();
     let secret_gen = DLogSecretGenService::init(key_gen_material, secret_manager_service);
     let key_id = OprfKeyId::from(U160::from(WRONG_ROUND_LOAD_PEER_PUBLIC_KEYS));
     let epoch = ShareEpoch::default();
@@ -393,8 +211,8 @@ async fn test_round2_in_wrong_round_during_load_public_keys() -> eyre::Result<()
 
 #[tokio::test]
 async fn test_abort() -> eyre::Result<()> {
-    let (base, secret_manager) = test_secret_manager();
-    let secret_manager_service: SecretManagerService = secret_manager.clone();
+    let secret_manager = test_secret_manager();
+    let secret_manager_service: SecretManagerService = secret_manager.service();
     let secret_gen = DLogSecretGenService::init(
         key_gen_material(DeploySetup::TwoThree),
         secret_manager_service,
@@ -403,7 +221,7 @@ async fn test_abort() -> eyre::Result<()> {
     let oprf_key_id = OprfKeyId::new(U160::from(142));
     let confirmed_epoch = ShareEpoch::default();
     let pending_epoch = confirmed_epoch.next();
-    base.add_random_key_material_with_id_epoch(
+    secret_manager.add_random_key_material_with_id_epoch(
         oprf_key_id,
         confirmed_epoch,
         &mut rand::thread_rng(),
@@ -458,8 +276,8 @@ async fn test_abort() -> eyre::Result<()> {
 
 #[tokio::test]
 async fn test_not_enough_producers() -> eyre::Result<()> {
-    let (base, secret_manager) = test_secret_manager();
-    let secret_manager_service: SecretManagerService = secret_manager.clone();
+    let secret_manager = test_secret_manager();
+    let secret_manager_service: SecretManagerService = secret_manager.service();
     let secret_gen = DLogSecretGenService::init(
         key_gen_material(DeploySetup::TwoThree),
         secret_manager_service,
@@ -468,7 +286,7 @@ async fn test_not_enough_producers() -> eyre::Result<()> {
     let oprf_key_id = OprfKeyId::new(U160::from(242));
     let confirmed_epoch = ShareEpoch::default();
     let pending_epoch = confirmed_epoch.next();
-    base.add_random_key_material_with_id_epoch(
+    secret_manager.add_random_key_material_with_id_epoch(
         oprf_key_id,
         confirmed_epoch,
         &mut rand::thread_rng(),
